move video track logic into TileVideo component
