+++ src/components/Tile/Tile.js
@@ -1,22 +1,13 @@
 import './Tile.css';
-import { useEffect, useMemo, useRef } from 'react';
+import { useEffect, useRef } from 'react';
 import { useMediaTrack } from '@daily-co/daily-react-hooks';
 import Username from '../Username/Username';
 import TileVideo from '../TileVideo/TileVideo';
 
 export default function Tile({ id, isScreenShare, isLocal, isAlone }) {
-  const videoTrack = useMediaTrack(id, isScreenShare ? 'screenVideo' : 'video');
   const audioTrack = useMediaTrack(id, isScreenShare ? 'screenAudio' : 'audio');
 
-  const videoElement = useRef(null);
   const audioElement = useRef(null);
-
-  useEffect(() => {
-    /*  The track is ready to be played. We can show video of the remote participant in the UI.*/
-    const video = videoElement.current;
-    if (!video || !videoTrack?.persistentTrack) return;
-    video.srcObject = new MediaStream([videoTrack?.persistentTrack]);
-  }, [videoTrack?.persistentTrack]);
 
   useEffect(() => {
     if (audioTrack?.state === 'playable') {
@@ -25,7 +16,6 @@ export default function Tile({ id, isScreenShare, isLocal, isAlone }) {
           audioTrack && new MediaStream([audioTrack.persistentTrack]));
     }
   }, [audioTrack]);
-
   let containerCssClasses = isScreenShare ? 'tile-screenshare' : 'tile-video';
 
   if (isLocal) {
@@ -37,7 +27,7 @@ export default function Tile({ id, isScreenShare, isLocal, isAlone }) {
 
   return (
     <div className={containerCssClasses}>
-      {videoTrack && <TileVideo videoElement={videoElement} />}
+      {<TileVideo id={id} isScreenShare={isScreenShare} />}
       {audioTrack && <audio autoPlay playsInline ref={audioElement} />}
       <Username id={id} isLocal={isLocal} />
     </div>
